anomaly-detection/requirements.txt
torch>=1.12.0
torchvision>=0.13.0
numpy>=1.21.0
opencv-python>=4.5.0
matplotlib>=3.5.0
scikit-learn>=1.1.0
tqdm>=4.64.0
PyYAML>=6.0
Pillow>=9.0.0
tensorboard>=2.8.0
wandb>=0.12.0
